+++ like/src/graphics/index.ts
@@ -16,6 +16,7 @@ export type {
   ShapeProps,
   DrawProps,
   PrintProps,
+  PolygonProps,
 } from "./graphics";
 
 export type {